+++ R/col_types.R
@@ -266,6 +266,7 @@ col_concise <- function(x) {
     c = col_character(),
     D = col_date(),
     d = col_double(),
+    f = col_factor(),
     i = col_integer(),
     l = col_logical(),
     n = col_number(),